refactor: remove deprecated mouth water plotting module
Deleted the mouth_water_plotter.py file as it is no longer needed. The functionality has been integrated into the updated plotting system, streamlining the codebase and improving maintainability.

Potential Signal for Alligator's Mouth state and water state #32
